+++ R/save_promo_image.R
@@ -30,22 +30,19 @@ save_promo_image <- function(game, gk = game_kit(), file = NULL) {
     envir <- list(piecepack = cfg)
     switch(game_name,
            nine_mens_morris = {
-               df <- promo_morris_df(game_name)
+               df <- promo_morris_df()
                wh <- render_piece(df, file = file, annotate = FALSE,
                                   envir = envir,
                                   op_scale = 0.5, trans = op_transform)
            },
            pass_the_food = {
-               df_tiles <- get_starting_df_from_name(game_name)
-               df_tiles$id <- NULL
-               df_tiles$cfg <- NULL
-               withr::local_seed(36)
-               df_coins <- tibble(piece_side = "coin_back",
-                                  rank = rep(1:6, 4), suit = rep(1:4, each=6),
-                                  x = stats::runif(24, -1.10, 8.70),
-                                  y = stats::runif(24, -1.20, 13.50),
-                                  angle = stats::runif(24, 0, 360))
-               df <- bind_rows(df_tiles, df_coins)
+               df <- promo_pass_the_food_df()
+               wh <- render_piece(df, file = file, annotate = FALSE,
+                                  envir = envir,
+                                  op_scale = 0.5, trans = op_transform)
+           },
+           speedy_towers = {
+               df <- promo_speedy_towers_df()
                wh <- render_piece(df, file = file, annotate = FALSE,
                                   envir = envir,
                                   op_scale = 0.5, trans = op_transform)
@@ -58,7 +55,7 @@ save_promo_image <- function(game, gk = game_kit(), file = NULL) {
 
            },
            twelve_mens_morris = {
-               df <- promo_morris_df(game_name)
+               df <- promo_morris_df()
                wh <- render_piece(df, file = file, annotate = FALSE,
                                   envir = envir,
                                   op_scale = 0.5, trans = op_transform)
@@ -72,12 +69,48 @@ save_promo_image <- function(game, gk = game_kit(), file = NULL) {
     invisible(c(wh, list(file = file)))
 }
 
-promo_morris_df <- function(game_name) {
-   df_tiles <- df_nine_mens_morris()
-   df_coins <- tibble(piece_side = "coin_back",
-                      x = c(3, 3, 5, 7, 7, 9, 9, 9, 11, 11, 11),
-                      y = c(3, 7, 7, 1, 3, 5, 7, 9, 3, 7, 11),
-                      suit = c(3, 4, 4, 1, 2, 2, 1, 1, 4, 3, 3),
-                      rank = c(1, 1, 2, 1, 1, 2, 2, 3, 3, 2, 3))
-   bind_rows(df_tiles, df_coins)
+promo_morris_df <- function() {
+    df_tiles <- df_nine_mens_morris()
+    df_coins <- tibble(piece_side = "coin_back",
+                       x = c(3, 3, 5, 7, 7, 9, 9, 9, 11, 11, 11),
+                       y = c(3, 7, 7, 1, 3, 5, 7, 9, 3, 7, 11),
+                       suit = c(3, 4, 4, 1, 2, 2, 1, 1, 4, 3, 3),
+                       rank = c(1, 1, 2, 1, 1, 2, 2, 3, 3, 2, 3))
+    bind_rows(df_tiles, df_coins)
+}
+
+promo_pass_the_food_df <- function() {
+    df_tiles <- df_pass_the_food()
+    withr::local_seed(36)
+    df_coins <- tibble(piece_side = "coin_back",
+                       rank = rep(1:6, 4), suit = rep(1:4, each=6),
+                       x = stats::runif(24, -1.10, 8.70),
+                       y = stats::runif(24, -1.20, 13.50),
+                       angle = stats::runif(24, 0, 360))
+    bind_rows(df_tiles, df_coins)
+}
+
+promo_speedy_towers_df <- function() {
+    withr::local_seed(72)
+    df_tiles <- tibble(piece_side = "tile_face",
+                       suit = rep(1:4, 6),
+                       rank = rep(1:6, each = 4))[sample.int(24L), ]
+    df_tiles$x <- c(5, rep(1, 10), rep(5, 12), 3)
+    df_tiles$y <- c(rep(1, 23), 7)
+    df_tiles$angle <- c(c(rep_len(0, 12), rep_len(180, 11))[sample.int(23L)], 180)
+
+    df_coins <- tibble(piece_side = "coin_back",
+                       suit = rep(1:4, 6),
+                       rank = rep(1:6, each = 4))[sample.int(24L), ]
+    df_coins$x <- c(rep(1, 10), rep(5, 12), 5, 6)
+    df_coins$y <- c(rep(1, 22), 7, 7)
+    df_coins$angle <- c(c(rep_len(0, 12), rep_len(180, 10))[sample.int(22L)], 180, 180)
+    # interleave coins
+    df_inter <- bind_rows(df_tiles[3:23, ], df_coins[1:22, ])[sample.int(43L), ]
+    # add pawns
+    df_pawns <- tibble(piece_side = "pawn_face",
+                       suit = c(1, 3), rank = NA_integer_,
+                       x = c(1, 5), y = c(7, 1), angle = c(180, 0))
+    df <- bind_rows(df_tiles[1:2, ], df_inter, df_pawns, df_tiles[24,, ], df_coins[23:24,, ])
+    df
 }